+++ src/main/java/executor/service/service/parallelflowexecutor/impls/ExecutorScenarioCreatorTask.java
@@ -1,4 +1,4 @@
-package executor.service.service.parallelflowexecutor.tasksimpl;
+package executor.service.service.parallelflowexecutor.impls;
 
 import executor.service.queue.ProxyQueue;
 import executor.service.queue.ScenarioQueue;
